moving dual tree related stuff to dual tree
