+++ hgeometry/hgeometry.cabal
@@ -367,6 +367,7 @@ library
         HGeometry.Polygon.Simple.Class
         HGeometry.Polygon.Simple.Sample
         HGeometry.Polygon.Simple.ShortestPath.Tree
+        HGeometry.Polygon.Simple.DualTree
 
 --
         HGeometry.Polygon.Convex
